backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
pandas
pdfplumber
openai
python-dotenv